flask/requirements.txt
flask
flask-smorest
python-dotenv
flask-sqlalchemy
mysql-connector-python
flask_socketio
flask_cors
numpy
tensorflow
pillow
matplotlib